move logic to superfluid
