--- proto/osmosis/gamm/v1beta1/unpool.proto
@@ -1,6 +1,0 @@
-syntax = "proto3";
-package osmosis.gamm.v1beta1;
-
-option go_package = "github.com/osmosis-labs/osmosis/v7/x/gamm/types";
-
-message UnpoolWhitelistedPools { repeated uint64 ids = 1; }
