+++ x/gamm/keeper/keeper.go
@@ -30,14 +30,13 @@ type Keeper struct {
 	hooks      types.GammHooks
 
 	// keepers
-	accountKeeper    types.AccountKeeper
-	bankKeeper       types.BankKeeper
-	distrKeeper      types.DistrKeeper
-	lockupKeeper     types.LockupKeeper
-	superfluidKeeper types.SuperfluidKeeper
+	accountKeeper types.AccountKeeper
+	bankKeeper    types.BankKeeper
+	distrKeeper   types.DistrKeeper
+	lockupKeeper  types.LockupKeeper
 }
 
-func NewKeeper(cdc codec.BinaryCodec, storeKey sdk.StoreKey, paramSpace paramtypes.Subspace, accountKeeper types.AccountKeeper, bankKeeper types.BankKeeper, distrKeeper types.DistrKeeper, lockupKeeper types.LockupKeeper, superfluidKeeper types.SuperfluidKeeper) Keeper {
+func NewKeeper(cdc codec.BinaryCodec, storeKey sdk.StoreKey, paramSpace paramtypes.Subspace, accountKeeper types.AccountKeeper, bankKeeper types.BankKeeper, distrKeeper types.DistrKeeper, lockupKeeper types.LockupKeeper) Keeper {
 	// Ensure that the module account are set.
 	moduleAddr, perms := accountKeeper.GetModuleAddressAndPermissions(types.ModuleName)
 	if moduleAddr == nil {
@@ -57,11 +56,10 @@ func NewKeeper(cdc codec.BinaryCodec, storeKey sdk.StoreKey, paramSpace paramtyp
 		cdc:        cdc,
 		paramSpace: paramSpace,
 		// keepers
-		accountKeeper:    accountKeeper,
-		bankKeeper:       bankKeeper,
-		distrKeeper:      distrKeeper,
-		lockupKeeper:     lockupKeeper,
-		superfluidKeeper: superfluidKeeper,
+		accountKeeper: accountKeeper,
+		bankKeeper:    bankKeeper,
+		distrKeeper:   distrKeeper,
+		lockupKeeper:  lockupKeeper,
 	}
 }
